Refuse loading executable files without executable code
HyperCPU emulator loads the executable file, calculates the size of executable
code in it (aka "binary size"), and then allocates the memory for reading and
executing the code. If the executable code is missing for some reason, emulator
will fail with std::bad_alloc exception.

This commit fixes this bug and adds a little check if the binary size is more
than zero.

Signed-off-by: Ivan Movchan <ivan.movchan.07@gmail.com>