Fix invalid CPU halt without HALT command
HyperCPU emulator runs the executable code until HALT command is sent.
If the end of the code is reached, and there is no HALT command,
emulator will continue reading and executing the code and finally abort
with error like "Interrupt was triggered, but failed to execute handler"
with XIP exceeding the size of the executable code (binary size).

This commit implements automatic CPU halt in case XIP exceeds the binary size.

Signed-off-by: Ivan Movchan <ivan.movchan.07@gmail.com>