2025-07-17 bugfixes (#38)
* Refuse loading executable files without executable code

HyperCPU emulator loads the executable file, calculates the size of executable
code in it (aka "binary size"), and then allocates the memory for reading and
executing the code. If the executable code is missing for some reason, emulator
will fail with std::bad_alloc exception.

This commit fixes this bug and adds a little check if the binary size is more
than zero.

Signed-off-by: Ivan Movchan <ivan.movchan.07@gmail.com>

* Fix invalid CPU halt without HALT command

HyperCPU emulator runs the executable code until HALT command is sent.
If the end of the code is reached, and there is no HALT command,
emulator will continue reading and executing the code and finally abort
with error like "Interrupt was triggered, but failed to execute handler"
with XIP exceeding the size of the executable code (binary size).

This commit implements automatic CPU halt in case XIP exceeds the binary size.

Signed-off-by: Ivan Movchan <ivan.movchan.07@gmail.com>

* Move XIP exception handling to interrupt handler

The previous solution in HyperCPU::CPU::Run() method could lead to performance
degradation.

Signed-off-by: Ivan Movchan <ivan.movchan.07@gmail.com>

* Improve binary file validation check

HyperCPU emulator will refuse binary file if its executable code size is
lesser than declared in the header, or if its header is broken.

Signed-off-by: Ivan Movchan <ivan.movchan.07@gmail.com>

* Fix CPU_TEST.INSTR_INTR_R_b64 test regression

Signed-off-by: Ivan Movchan <ivan.movchan.07@gmail.com>

* Fix EXCEPTION_TEST.CPU_EXCEPTION_IO test

Signed-off-by: Ivan Movchan <ivan.movchan.07@gmail.com>

* Improved binary file verification

Most of the code responsible for binary file verification is now available
in VerifyBinaryFile method.

The emulator now refuses to run the binary file with a size less than default
binary header size.

Signed-off-by: Ivan Movchan <ivan.movchan.07@gmail.com>

* Improve binary file verification and fix a little bug

I wrongly typed !VerifyBinaryFile instead of VerifyBinaryFile. As a result,
emulator could continue working in case VerifyBinaryFile returned 1 (error)
or shutdown if 0 (no errors).

Signed-off-by: Ivan Movchan <ivan.movchan.07@gmail.com>

---------

Signed-off-by: Ivan Movchan <ivan.movchan.07@gmail.com>